+++ app/src/main/java/com/pavlov/nearWarSecrets/ui/Images/MemeSelectionDialog.kt
@@ -1,38 +1,42 @@
-package com.pavlov.nearWarSecrets.ui.itemLoader
+package com.pavlov.nearWarSecrets.ui.Images
 
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.runtime.Composable
-import android.content.Intent
-import android.net.Uri
-import android.widget.Toast
+import com.pavlov.nearWarSecrets.R
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.window.Dialog
-import coil.compose.rememberImagePainter
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.graphics.Color
 
 @Composable
-fun ExtractedImagesDialog(
-    extractedImages: List<Uri>,
+fun MemeSelectionDialog(
+    onMemeSelected: (Int) -> Unit,
     onDismiss: () -> Unit
 ) {
+    val memes = listOf(
+        R.drawable.mem1,
+        R.drawable.mem2,
+        R.drawable.mem3
+    )
+
     Dialog(onDismissRequest = onDismiss) {
         Surface(
             shape = RoundedCornerShape(8.dp),
@@ -41,37 +45,31 @@ fun ExtractedImagesDialog(
                 .border(width = 2.dp, color = Color.Green, shape = RoundedCornerShape(8.dp))
         ) {
             Column(
-                modifier = Modifier
-                    .padding(16.dp)
-                    .fillMaxWidth()
+                modifier = Modifier.padding(16.dp)
             ) {
-                Text(
-                    text = "Извлеченные изображения",
-                    style = MaterialTheme.typography.h6
-                )
+                Text(text = "Выберите мем", style = MaterialTheme.typography.h6)
                 Spacer(modifier = Modifier.height(8.dp))
-                LazyColumn(
+                LazyRow(
+                    horizontalArrangement = Arrangement.spacedBy(8.dp),
                     modifier = Modifier.fillMaxWidth()
                 ) {
-                    // Используем правильный метод для работы с коллекцией
-                    items(extractedImages) { uri ->
+                    items(items = memes) { memeResId ->
                         Image(
-                            painter = rememberImagePainter(uri),
+                            painter = painterResource(id = memeResId),
                             contentDescription = null,
                             modifier = Modifier
-                                .fillMaxWidth()
-                                .height(200.dp)
-                                .padding(vertical = 4.dp)
+                                .size(100.dp)
+                                .clickable {
+                                    onMemeSelected(memeResId)
+                                    onDismiss()
+                                }
                                 .clip(RoundedCornerShape(8.dp))
                         )
                     }
                 }
                 Spacer(modifier = Modifier.height(8.dp))
-                Button(
-                    onClick = onDismiss,
-                    modifier = Modifier.align(Alignment.End)
-                ) {
-                    Text("Закрыть")
+                Button(onClick = onDismiss, modifier = Modifier.align(Alignment.End)) {
+                    Text("Отмена")
                 }
             }
         }
