+++ app/src/main/java/com/pavlov/nearWarSecrets/ui/Images/saveImageToGallery.kt
@@ -1,4 +1,4 @@
-package com.pavlov.nearWarSecrets.ui.itemLoader
+package com.pavlov.nearWarSecrets.ui.Images
 
 import android.content.Context
 import android.content.Intent
